+++ lib/src/framework/provide_it_scope.dart
@@ -2,9 +2,7 @@ part of '../framework.dart';
 
 class ProvideItScope implements ReadIt {
   static ProvideItScope of(BuildContext context) {
-    final it = context.getElementForInheritedWidgetOfExactType<ProvideIt>();
-    assert(it != null, 'You must set a `ProvideIt` above your app.');
-    return (it as ProvideItElement).scope;
+    return ProvideItElement.of(context).scope;
   }
 
   /// The attached [ProvideIt] element.
@@ -59,43 +57,35 @@ class ProvideItScope implements ReadIt {
   bool get mounted => _element != null;
 
   // bind tree by context and index.
-  late final _tree = TreeMap<Element, TreeMap<int, Bind>>()._assert(this);
-  late final _treeIndex = <Element, int>{}._assert(this);
+  // manages the lifecycle of the binds.
+  late final _binds = TreeMap<Element, TreeMap<int, Bind>>()._assert(this);
+  late final _bindIndex = <Element, int>{}._assert(this);
 
-  // tree cache
-  late final _typeCache = <String, Set<Bind>>{}._assert(this);
-  late final _refCache = <Ref, Set<Bind>>{}._assert(this);
+  // bind tree cache.
+  // used to find the bind by type.
+  late final _bindCache = <String, Set<Bind>>{}._assert(this);
 
-  // bind dependencies by the dependent `context`.
-  late final _dependencies = <Element, Set<Bind>>{}._assert(this);
-  late final _dependencyIndex = <Element, int>{}._assert(this);
+  // bind observers by the dependent `context`.
+  // manages which observers can be notified.
+  late final _observers = <Element, Set<Bind>>{}._assert(this);
+  late final _observerIndex = <Element, int>{}._assert(this);
 
   /// Iterates over all [Ref] binds. Depth-first.
   Iterable<Bind> get binds sync* {
-    for (var branch in _tree.values) {
+    for (var branch in _binds.values) {
       for (var bind in branch.values) {
-        yield bind;
+        if (!bind.deactivated) yield bind;
       }
     }
   }
 
   void _register(Bind bind) {
-    final binds = switch (bind.key) {
-      Ref ref => _refCache[ref] ??= {},
-      _ => _typeCache[bind.type] ??= {},
-    };
+    final binds = _bindCache[bind.type] ??= {};
     binds.add(bind);
   }
 
   void _unregister(Bind bind) {
-    final (cache, key) = switch (bind.key) {
-      Ref ref => (_refCache, ref),
-      _ => (_typeCache, bind.type),
-    };
-    final binds = cache[key]!;
-    binds.remove(bind);
-
-    if (binds.isEmpty) cache.remove(key);
+    _bindCache[bind.type]!.remove(bind);
   }
 
   @protected
@@ -106,15 +96,22 @@ class ProvideItScope implements ReadIt {
   /// The future of [AsyncBind.isReady].
   @override
   FutureOr<void> allReady() {
-    final futures = <Future>[];
+    final futures = <Future>{};
 
     for (var bind in binds) {
       if (bind is! AsyncBind) continue;
-      if (bind.isReady() case Future it) futures.add(it);
+      if (bind.isReady() case Future it) {
+        futures.add(it.onError((e, s) {
+          Error.throwWithStackTrace(
+            '#${bind.index} ${bind.type} $e',
+            s,
+          );
+        }));
+      }
     }
     if (futures.isEmpty) return null;
 
-    return futures.wait as Future<void>;
+    return Future.wait(futures, eagerError: true) as Future<void>;
   }
 
   /// The future when a [AsyncBind.isReady] is completed.
@@ -123,18 +120,17 @@ class ProvideItScope implements ReadIt {
     type ??= T.type;
 
     final binds = switch (key) {
-      Ref ref => _refCache[ref],
-      null => _typeCache[type],
-      _ => _typeCache[type]?.where((it) => it.key == key),
+      null => _bindCache[type] ?? {},
+      _ => _bindCache[type]?.where((it) => it.key == key) ?? {},
     };
 
-    assert(binds != null, 'AsyncRef<$type> not found, key: $key.');
+    assert(binds.isNotEmpty, 'provide<$type> not found, key: $key.');
     assert(
-      binds?.length == 1 || binds?.where((it) => !it.deactivated).length == 1,
-      'Duplicate AsyncRef<$type>, key: $key.',
+      binds.length == 1 || binds.where((it) => !it.deactivated).length == 1,
+      'Duplicate provide<$type>, key: $key.',
     );
 
-    if (binds?.lastOrNull case AsyncBind bind) return bind.isReady();
+    if (binds.lastOrNull case AsyncBind bind) return bind.isReady();
     return null;
   }
 
@@ -164,7 +160,7 @@ class ProvideItScope implements ReadIt {
     if (value == null && null is T) return value;
     if (bind != null) return bind.read();
 
-    throw StateError('Ref<$T> not found, key: $key.');
+    throw MissingProvideException('$T not found.');
   }
 
   @override
@@ -190,7 +186,8 @@ Did you provide the missing type?
 context.provide<$type>(...); // <- provide it
 ''',
     );
-    throw StateError('AsyncRef<$type> not found, key: $key.');
+
+    throw MissingProvideException('$type not found.');
   }
 
   @protected
@@ -198,21 +195,20 @@ context.provide<$type>(...); // <- provide it
     type ??= T.type;
 
     final binds = switch (key) {
-      Ref ref => _refCache[ref] ?? {},
-      null => _typeCache[type] ?? {},
-      _ => _typeCache[type]?.where((b) => Ref.equals(b.key, key)) ?? {},
+      null => _bindCache[type] ?? {},
+      _ => _bindCache[type]?.where((b) => Ref.equals(b.key, key)) ?? {},
     };
     final bind = binds.lastOrNull;
 
     assert(
       binds.where((it) => !it.deactivated).length < 2,
-      'Duplicate Ref<$type> found, key: $key.',
+      'Duplicate provide<$type> found, key: $key.',
     );
     return bind;
   }
 
   @override
-  String toString() => _tree.toString();
+  String toString() => _binds.toString();
 }
 
 extension<K, V> on Map<K, V> {
@@ -263,4 +259,20 @@ class AssertMap<K, V> extends MapBase<K, V> {
     _assert();
     return _map[key];
   }
+}
+
+class MissingProvideException implements Exception {
+  MissingProvideException(this.message);
+  final String message;
+
+  @override
+  String toString() => 'MissingProvideException: $message';
+}
+
+class LoadingProvideException implements Exception {
+  LoadingProvideException(this.message);
+  final String message;
+
+  @override
+  String toString() => 'LoadingProvideException: $message';
 }