+++ lib/src/refs/ref.dart
@@ -4,6 +4,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 
 import '../../provide_it.dart';
+import '../framework.dart';
 
 /// The base class for all [Ref] types.
 ///
@@ -19,7 +20,7 @@ import '../../provide_it.dart';
 ///
 /// See [Bind] for more details.
 abstract class Ref<T> {
-  const Ref({this.key = id});
+  const Ref({this.key});
 
   /// The unique identifier for the [Ref] instance.
   /// Similar to [Widget.key].
@@ -39,18 +40,13 @@ abstract class Ref<T> {
       };
 
   /// A [Widget.canUpdate] implementation for [Ref] with [equals].
-  static bool canUpdate(Ref oldRef, Ref newRef) =>
-      identical(oldRef, newRef) ||
-      (oldRef.runtimeType == newRef.runtimeType &&
-          equals(oldRef.key, newRef.key));
+  static bool canUpdate(Ref oldRef, Ref newRef) {
+    return oldRef.runtimeType == newRef.runtimeType &&
+        equals(oldRef.key, newRef.key);
+  }
 
   /// The equality used for [Ref.key] & [Bind.select].
   static bool equals(Object? a, Object? b) => defaultEquals(a, b);
-
-  /// Signature for using [Ref] own identity as [key].
-  /// Only available for a top-level [Ref].
-  @protected
-  static void id() {}
 
   @protected
   Bind<T, Ref<T>> createBind();
@@ -59,76 +55,7 @@ abstract class Ref<T> {
 extension RefReaders<T> on Ref<T> {
   /// Binds this [Ref] to the [BuildContext].
   @protected
-  Bind<T, Ref<T>> bind(BuildContext context) => context.bind(this);
-
-  /// Gets the [Bind] of this [Ref].
-  @protected
-  Bind<T, Ref<T>> bindOf(BuildContext context) {
-    return context.bindOf<T>(key: this) as Bind<T, Ref<T>>;
+  Bind<T, Ref<T>> bind(BuildContext context) {
+    return ProvideItScope.of(context).bind(context, this);
   }
-
-  /// Reads a previously bound [T] value.
-  T read(BuildContext context) {
-    return context.read(key: this);
-  }
-
-  /// Watches this [T] value. Auto-binds.
-  T watch(BuildContext context) {
-    return context.watch(key: this);
-  }
-
-  /// Selects this [R] value using [selector]. Auto-binds.
-  R select<R>(BuildContext context, R selector(T value)) {
-    return context.select(selector, key: this);
-  }
-
-  /// Listens this [T] value using [listener]. Auto-binds.
-  void listen(BuildContext context, void listener(T value)) {
-    context.listen(listener, key: this);
-  }
-
-  /// Listens this [R] value using [selector] and [listener]. Auto-binds.
-  void listenSelect<R>(
-    BuildContext context,
-    R selector(T value),
-    void listener(R? previous, R next),
-  ) {
-    context.listenSelect(selector, listener, key: this);
-  }
-}
-
-extension AsyncRefBinder<T> on AsyncRef<T> {
-  /// Binds the [AsyncRef] to the [BuildContext].
-  AsyncBind<T, AsyncRef<T>> bind(BuildContext context) {
-    return context.bind(this) as AsyncBind<T, AsyncRef<T>>;
-  }
-
-  AsyncBind<T, AsyncRef<T>> bindOf(BuildContext context) {
-    return context.bindOf<T>(key: this) as AsyncBind<T, AsyncRef<T>>;
-  }
-
-  /// Watches the [AsyncSnapshot] of this async value.
-  AsyncSnapshot<T> watch(BuildContext context) {
-    return bindOf(context).watch(context);
-  }
-}
-
-extension AsyncRefReaders<T> on AsyncRef<T> {
-  /// Reloads the value of this [Ref].
-  Future<void> reload(BuildContext context) {
-    return context.reload<T>(key: this);
-  }
-
-  /// Async reads the value of this [Ref].
-  FutureOr<T> readAsync(BuildContext context) {
-    return context.readAsync<T>(key: this);
-  }
-
-  /// The future when this [Ref] is ready to be read.
-  FutureOr<void> isReady(BuildContext context) {
-    return context.readAsync<T>(key: this);
-  }
-
-  /// Whether this [Ref] is ready to be read.
-  bool isReadySync(BuildContext context) => isReady(context) == null;
 }